+++ SamplesWeightingApp/src/models/SRMs.cs
@@ -8,8 +8,8 @@ namespace SamplesWeighting
         public string SRM_Set_Name     { get; set; }
         public string SRM_Set_Number   { get; set; }
         public string SRM_Number       { get; set; }
-        public decimal? SRM_SLI_Weight { get; set; }
-        public decimal? SRM_LLI_Weight { get; set; }
+        public float? SRM_SLI_Weight { get; set; }
+        public float? SRM_LLI_Weight { get; set; }
     }
 
     [Table("table_SRM_Set")]
